Rename dds subscriber file
