+++ dds/src/implementation/dds_impl/mod.rs
@@ -28,4 +28,4 @@ pub mod reader_factory;
 pub mod status_condition_impl;
 pub mod status_listener;
 pub mod topic_impl;
-pub mod user_defined_subscriber;
+pub mod dds_subscriber;
